// src/interfaces/dictamen.interface.ts
export interface Dictamen {
    uuid?: string;
    numero_oficio?: string;
    fecha_oficio?: string;
    pase_caja?: string;
    numero_expediente?: string
    folio_electronico?: string;
    fecha_impresion?: string;
    vigencia?: string;
    licencia_funcionamiento?: number
    nombre_comercial?: string
    razon_social?: string
    giro?: string
    direccion?: string
    rfc?: string
    actividad?: string

}